Remove default exports; less trivial cases in top-level runtime dir.
