+++ runtime/api-channel.js
@@ -11,7 +11,7 @@
 
 import assert from '../platform/assert-web.js';
 import {ParticleSpec} from './particle-spec.js';
-import Type from './type.js';
+import {Type} from './type.js';
 import OuterPortAttachment from './debug/outer-port-attachment.js';
 
 class ThingMapper {
