+++ runtime/arc.js
@@ -9,11 +9,10 @@
  */
 'use strict';
 
-import runtime from './runtime.js';
 import assert from '../platform/assert-web.js';
-import Type from './type.js';
+import {Type} from './type.js';
 import {handleFor} from './handle.js';
-import OuterPec from './outer-PEC.js';
+import {OuterPEC} from './outer-PEC.js';
 import Recipe from './recipe/recipe.js';
 import {Manifest} from './manifest.js';
 import {Description} from './description.js';
@@ -53,7 +52,7 @@ class Arc {
     this.particleHandleMaps = new Map();
     let pecId = this.generateID();
     let innerPecPort = this._pecFactory(pecId);
-    this.pec = new OuterPec(innerPecPort, slotComposer, this, `${pecId}:outer`);
+    this.pec = new OuterPEC(innerPecPort, slotComposer, this, `${pecId}:outer`);
     if (slotComposer) {
       slotComposer.arc = this;
     }
