+++ runtime/dom-particle.js
@@ -10,11 +10,7 @@
 'use strict';
 
 import assert from '../platform/assert-web.js';
-import {
-  Particle,
-  ViewChanges
-} from './particle.js';
-
+import {Particle, ViewChanges} from './particle.js';
 import XenStateMixin from '../shell/components/xen/xen-state.js';
 
 /** @class DomParticle
