+++ runtime/inner-PEC.js
@@ -9,7 +9,7 @@
  */
 'use strict';
 
-import Type from './type.js';
+import {Type} from './type.js';
 import {handleFor} from './handle.js';
 import assert from '../platform/assert-web.js';
 import {PECInnerPort} from './api-channel.js';
@@ -255,4 +255,4 @@ class InnerPEC {
   }
 }
 
-export default InnerPEC;
+export {InnerPEC};
