+++ runtime/manual_tests/firebase-tests.js
@@ -11,7 +11,7 @@
 import StorageProviderFactory from '../storage/storage-provider-factory.js';
 import {Arc} from '../arc.js';
 import {Manifest} from '../manifest.js';
-import Type from '../type.js';
+import {Type} from '../type.js';
 import assert from '../../platform/assert-web.js';
 
 describe('firebase', function() {
