+++ runtime/particle-spec.js
@@ -8,7 +8,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import Type from './type.js';
+import {Type} from './type.js';
 import TypeChecker from './recipe/type-checker.js';
 import {Shape} from './shape.js';
 import assert from '../platform/assert-web.js';
