+++ runtime/planificator.js
@@ -6,7 +6,7 @@
 // http://polymer.github.io/PATENTS.txt
 
 import assert from '../platform/assert-web.js';
-import Type from './type.js';
+import {Type} from './type.js';
 import InitSearch from './strategies/init-search.js';
 import {Planner} from './planner.js';
 import {Speculator} from './speculator.js';
