+++ runtime/relation.js
@@ -9,7 +9,7 @@
 
 import assert from '../platform/assert-web.js';
 import {Entity} from './entity.js';
-import Type from './type.js';
+import {Type} from './type.js';
 import Symbols from './symbols.js';
 
 // TODO: Should relations normalized by another layer, or here?
