+++ runtime/test/particle-shape-loading-test.js
@@ -13,10 +13,10 @@ import {assert} from './chai-web.js';
 import * as util from './test-util.js';
 import {Arc} from '../arc.js';
 import {MessageChannel} from '../message-channel.js';
-import InnerPec from '../inner-PEC.js';
+import {InnerPEC} from '../inner-PEC.js';
 import {Loader} from '../loader.js';
 import Recipe from '../recipe/recipe.js';
-import Type from '../type.js';
+import {Type} from '../type.js';
 import {Shape} from '../shape.js';
 import {ParticleSpec} from '../particle-spec.js';
 
@@ -72,7 +72,7 @@ describe('particle-shape-loading', function() {
 
     let pecFactory = function(id) {
       let channel = new MessageChannel();
-      new InnerPec(channel.port1, `${id}:inner`, loader);
+      new InnerPEC(channel.port1, `${id}:inner`, loader);
       return channel.port2;
     };
 
@@ -137,7 +137,7 @@ describe('particle-shape-loading', function() {
 
     let pecFactory = function(id) {
       let channel = new MessageChannel();
-      new InnerPec(channel.port1, `${id}:inner`, loader);
+      new InnerPEC(channel.port1, `${id}:inner`, loader);
       return channel.port2;
     };
