+++ runtime/test/particle-shape-loading-with-slots-test.js
@@ -14,10 +14,10 @@ import * as util from './test-util.js';
 import {handleFor} from '../handle.js';
 import {Arc} from '../arc.js';
 import {MessageChannel} from '../message-channel.js';
-import InnerPec from '../inner-PEC.js';
+import {InnerPEC} from '../inner-PEC.js';
 import {Loader} from '../loader.js';
 import Recipe from '../recipe/recipe.js';
-import Type from '../type.js';
+import {Type} from '../type.js';
 import MockSlotComposer from './mock-slot-composer.js';
 
 describe('particle-shape-loading-with-slots', function() {
@@ -25,7 +25,7 @@ describe('particle-shape-loading-with-slots', function() {
     let loader = new Loader();
     let pecFactory = function(id) {
       let channel = new MessageChannel();
-      new InnerPec(channel.port1, `${id}:inner`, loader);
+      new InnerPEC(channel.port1, `${id}:inner`, loader);
       return channel.port2;
     };
     let slotComposer = new MockSlotComposer();
