+++ shell/source/worker-entry.js
@@ -6,8 +6,8 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import InnerPec from '../../runtime/inner-PEC.js';
-import Loader from './browser-loader.js';
+import {InnerPEC} from '../../runtime/inner-PEC.js';
+import {BrowserLoader} from './browser-loader.js';
 
 const log = console.log.bind(console, `%cworker-entry`, `background: #12005e; color: white; padding: 1px 6px 2px 7px; border-radius: 6px;`);
 
@@ -15,5 +15,5 @@ self.onmessage = function(e) {
   self.onmessage = null;
   let {id, base} = e.data;
   //log('starting worker', id);
-  new InnerPec(e.ports[0], id, new Loader(base));
+  new InnerPEC(e.ports[0], id, new BrowserLoader(base));
 };
